pua: solved warning on comparing unsigned expression
(cherry picked from commit 55b42c5ba9e78b1d6662d45f1b68898bc67548c9)

+++ modules/pua/pua.c
@@ -993,7 +993,7 @@ static void db_update(unsigned int ticks,void *param)
 		p = HashT->p_records[i].entity->next;
 		while(p)
 		{
-			if(p->expires - (int)time(NULL)< 0)	
+			if((int)p->expires - (int)time(NULL)< 0)	
 			{
 				p= p->next;
 				continue;
